Un-done changes to unit tests for build tests.

+++ src/CommunityToolkit.Maui.UnitTests/Layouts/StateContainerTests.cs
@@ -219,7 +219,7 @@ public class StateContainerTests : BaseTest
 		await Assert.ThrowsAsync<TaskCanceledException>(() => StateContainer.ChangeStateWithAnimation(layout, StateKey.Error, cancelledTokenSource.Token));
 	}
 
-	[Fact(Timeout = (int)TestDuration.Short)]
+	[Fact(Timeout = (int)TestDuration.Long)]
 	public async Task StateContainer_CustomAnimation_Timeout()
 	{
 		layout.EnableAnimations();
@@ -284,7 +284,7 @@ public class StateContainerTests : BaseTest
 		Assert.True(StateContainer.GetCanStateChange(layout));
 	}
 
-	[Fact(Timeout = (int)TestDuration.Short)]
+	[Fact(Timeout = (int)TestDuration.Long)]
 	public async Task StateContainer_ChangingStateWhenCanStateChangePropertyIsFalse_CustomBeforeAnimation()
 	{
 		layout.EnableAnimations();
@@ -348,7 +348,7 @@ public class StateContainerTests : BaseTest
 		Assert.True(StateContainer.GetCanStateChange(layout));
 	}
 
-	[Fact(Timeout = (int)TestDuration.Short)]
+	[Fact(Timeout = (int)TestDuration.Long)]
 	public async Task StateContainer_ChangingStateWhenCanStateChangePropertyIsFalse_CustomBeforeAndAfterAnimation()
 	{
 		layout.EnableAnimations();
@@ -403,7 +403,7 @@ public class StateContainerTests : BaseTest
 		Assert.True(StateContainer.GetCanStateChange(layout));
 	}
 
-	[Fact(Timeout = (int)TestDuration.Short)]
+	[Fact(Timeout = (int)TestDuration.Long)]
 	public async Task StateContainer_ChangingStateWhenCanStateChangePropertyIsFalse_CustomBeforeAnimationFuncs()
 	{
 		layout.EnableAnimations();
@@ -461,7 +461,7 @@ public class StateContainerTests : BaseTest
 		static Task CustomAnimation(VisualElement element, CancellationToken token) => element.RotateTo(0.75, 500).WaitAsync(token);
 	}
 
-	[Fact(Timeout = (int)TestDuration.Short)]
+	[Fact(Timeout = (int)TestDuration.Long)]
 	public async Task StateContainer_ChangingStateWhenCanStateChangePropertyIsFalse_CustomBeforeAndAfterAnimationFuncs()
 	{
 		layout.EnableAnimations();
